refactor: Router owns the route table, modules own their routes
Router.handle was a 250-line switch with the workspace, zone, agent and
update handlers written out inside it, while /shot and /ruler already had
files of their own. Now every case is one line to the module that owns
it, and the switch reads as the list of what the app answers.

Router.swift 575 -> 183 lines, so it comes out of the line-limit baseline
rather than being lowered in it. Nothing moved between modules: the
handlers are the same code in Router+Workspaces, Router+Zones,
Router+Agents, Router+Update, Router+State and Router+Access.

Two duplications went with it:
- HTTPResponse.conflict replaces six hand-built 409s. A refusal now reads
  the same as .notFound and .badRequest beside it.
- The two /update routes shared a "the user turned checks off" refusal
  spelled out twice, which is a promise only kept if both keep it. One
  function now, called by both.

AGENTS.md follows: the module recipe names the Router+<Module> file and
the manager's apply, and says plainly that a setting is not an AppActions
method.
