+++ scripts/line-limit-baseline
@@ -9,7 +9,6 @@
 #
 # lines  path
 867   App/Sources/plonk/AppDelegate.swift
-575   App/Sources/plonk/Router.swift
 489   App/Tests/plonkTests/ConfigTests.swift
 444   App/Tests/plonkTests/RouterTests.swift
 428   App/Sources/plonk/WindowManager.swift
